fix(DASH/HLS): Don't merge folders, skip final merge if only 1 segment

+++ devine/core/manifests/dash.py
@@ -486,28 +486,33 @@ class DASH:
         for control_file in save_dir.glob("*.aria2__temp"):
             control_file.unlink()
 
-        segments_to_merge = sorted(save_dir.iterdir())
-        progress(downloaded="Merging", completed=0, total=len(segments_to_merge))
-
+        segments_to_merge = [
+            x
+            for x in sorted(save_dir.iterdir())
+            if x.is_file()
+        ]
         with open(save_path, "wb") as f:
             if init_data:
                 f.write(init_data)
-            for segment_file in segments_to_merge:
-                segment_data = segment_file.read_bytes()
-                # TODO: fix encoding after decryption?
-                if (
-                    not drm and isinstance(track, Subtitle) and
-                    track.codec not in (Subtitle.Codec.fVTT, Subtitle.Codec.fTTML)
-                ):
-                    segment_data = try_ensure_utf8(segment_data)
-                    segment_data = segment_data.decode("utf8"). \
-                        replace("&lrm;", html.unescape("&lrm;")). \
-                        replace("&rlm;", html.unescape("&rlm;")). \
-                        encode("utf8")
-                f.write(segment_data)
-                f.flush()
-                segment_file.unlink()
-                progress(advance=1)
+            if len(segments_to_merge) > 1:
+                progress(downloaded="Merging", completed=0, total=len(segments_to_merge))
+            else:
+                for segment_file in segments_to_merge:
+                    segment_data = segment_file.read_bytes()
+                    # TODO: fix encoding after decryption?
+                    if (
+                        not drm and isinstance(track, Subtitle) and
+                        track.codec not in (Subtitle.Codec.fVTT, Subtitle.Codec.fTTML)
+                    ):
+                        segment_data = try_ensure_utf8(segment_data)
+                        segment_data = segment_data.decode("utf8"). \
+                            replace("&lrm;", html.unescape("&lrm;")). \
+                            replace("&rlm;", html.unescape("&rlm;")). \
+                            encode("utf8")
+                    f.write(segment_data)
+                    f.flush()
+                    segment_file.unlink()
+                    progress(advance=1)
 
         track.path = save_path
         if callable(track.OnDownloaded):
